Large movies list refactoring
