+++ client.rest
@@ -1,0 +1,71 @@
+GET http://localhost:3001/crash-test
+
+###
+GET http://localhost:3001/health-check
+
+###
+
+GET http://localhost:3001/undefined
+
+###
+
+POST http://localhost:3001/signup
+Content-Type: application/json
+
+{
+    "email":"hho.edst@mail.com",
+    "name":"Pedro",
+    "password":"hereweare"
+}
+###
+POST http://localhost:3001/signin
+Content-Type: application/json
+
+{
+    "email":"hho.edst@mail.com",
+    "password":"hereweare"
+}
+###
+
+GET http://localhost:3001/users/me
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MTgwNDk2YmM4YzlkNDlmOGVhOGI0MjMiLCJpYXQiOjE2MzU3OTgxOTUsImV4cCI6MTYzNjQwMjk5NX0.xYOhfleEB_sIEVzk-VCUoaSlERfEzOr395x7HDuh0I0
+###
+
+PATCH http://localhost:3001/users/me
+Content-Type: application/json
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MTRkOWU0NTBjOTcwMGJiZTI4NGJkODgiLCJpYXQiOjE2MzI0NzY3NTIsImV4cCI6MTYzMzA4MTU1Mn0.QgkbhMMpdG9XT9a_07DvgAtq2e4mIkS4ycpP4RtiAZI
+
+{
+  "name": "Test one",
+  "email": "daisddemail@mail.com"
+}
+
+###
+GET http://localhost:3001/movies
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MTgwNDk2YmM4YzlkNDlmOGVhOGI0MjMiLCJpYXQiOjE2MzU4MDI2NDAsImV4cCI6MTYzNjQwNzQ0MH0.M1JC9C0he8cZCiFdBpn4k-jqAw41imWrsbc424vFnBo
+
+###
+
+
+POST http://localhost:3001/movies
+Content-Type: application/json
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MTgwNDk2YmM4YzlkNDlmOGVhOGI0MjMiLCJpYXQiOjE2MzU4MDI2NDAsImV4cCI6MTYzNjQwNzQ0MH0.M1JC9C0he8cZCiFdBpn4k-jqAw41imWrsbc424vFnBo
+
+{
+    "country":null,
+    "director": "Main one",
+    "duration": 123,
+    "year": "1233",
+    "description": "There is description",
+    "image": "https://images.mesto.ru/image.png",
+    "trailer": "https://images.mesto.ru/video.png",
+    "nameRU": "Фильм",
+    "nameEN": "Movie",
+    "thumbnail": "https://images.mesto.ru/image.png",
+    "movieId":123
+}
+
+###
+DELETE http://localhost:3001/movies/6149dde0b44d57b3dc688f66
+# Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MTQ5YThiODBlNzRhZTQ5MDI5MzUxNTIiLCJpYXQiOjE2MzIyMzA1NTMsImV4cCI6MTYzMjgzNTM1M30.xcn0XRPzJAFHah_NKsP1_qmfiEeF0xF9-UYoF97OObs
+
