+++ src/components/MoviesList/MoviesList.css
@@ -1,0 +1,38 @@
+.cards {
+    padding: 70px 70px 10px;
+  }
+  
+  .card-list {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+    display: grid;
+    gap: 24px;
+    grid-template-columns: repeat(auto-fit, minmax(364px, 1fr));
+    justify-items: center;
+  }
+  
+  @media screen and (max-width: 1279px) {
+    .cards {
+      padding: 70px 30px 30px;
+    }
+  
+    .card-list {
+      padding: 0;
+      gap: 30px;
+      grid-template-columns: repeat(auto-fit, minmax(339px, 1fr));
+    }
+  }
+  
+  @media screen and (max-width: 767px) {
+    .cards {
+      padding: 40px 10px 26px;
+    }
+  
+    .card-list {
+      padding: 0;
+      gap: 14px;
+      grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    }
+  }
+
